LearnJava

+++ Main.java
@@ -2,7 +2,7 @@
 public class Main{
     public static void main(String[] args){
 
-        int x = 123000000000000; //initialization
+        long x = 123000000000000; //initialization
 
         System.out.println("My number is: "+x);
     }
